image-api/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
torch>=2.0.0
diffusers>=0.30.0
transformers>=4.40.0
accelerate>=0.27.0
Pillow>=10.0.0
